Q1 removeNode()
. Implement the removeNode() function for a linked list, using the lecture
diagrams and pseudo-code as a reference. The prototype of the removeNode()
function is given below:
int removeNode(ListNode **ptrHead, int index);
The function should return 1 if the delete operation is successful and 0 otherwise.
Recall that the function requires a pointer to the head pointer in order to
correctly delete the first node.
Write a program to test the removeNode() function. It should first allow the
user to create a linked list of integers by appending values to the end of the list.
Next, it should allow the user to delete nodes one by one based on their indices.
The program can be terminated by providing an out-of-range index to delete the
node.
Sample inputs and outputs:
Enter a list of numbers, terminated by any non-digit character:
1
2
3
4
5
a
Current List: 1 2 3 4 5
Enter the index of the node to be removed: 0
After the removal operation,
Current List: 2 3 4 5
Enter the index of the node to be removed: 3
After the removal operation,
Current List: 2 3 4
Enter the index of the node to be removed: 1
After the removal operation,
Current List: 2 4
Enter the index of the node to be removed: 3
The node cannot be removed.
Current List: 2 4