Q1 palindrome()
Write a function palindrome() that determines whether a given string in a
Queue is a palindrome. The prototype of the palindrome() function is
given below:
int palindrome(Queue* wordPtr);
The function should return 1 if the string is a palindrome and 0 otherwise. You
should ignore the space characters and the case of each letter. The stack and
queue data structures and their interface functions are provided.
Sample inputs and outputs:
Enter a string of characters, terminated by a newline:
On A Clover If Alive ERUPTS A VAST PURE EVIL A FIRE Volcano
The string is a palindrome.
Enter a string of characters, terminated by a newline:
A man a plan a CANAL PANama
The string is a palindrome.
Enter a string of characters, terminated by a newline:
Superman in the sky
The string is not a palindrome.