Q2 balanced()
Write a function balanced() that determines if an expression comprised of
the characters, ()[]{}, is balanced. The prototype of the balanced()
function is given below:
int balanced(char *expression);
The function returns 1 if the expression is balanced. Otherwise, it return 0. The
following expressions are balanced because the order and quantity of the
parentheses match:
()
([])
{[]()[]}
The following expressions are not balanced:
{{)]
[({{)])
Sample inputs and outputs:
Enter an expression, terminated by a newline:
[({{{}}})[[]]{}({})]
The expression is balanced.
Enter an expression, terminated by a newline:
{1+2+{5}*[6+x]+{4+5}(3+2)}
The expression is balanced.
Enter an expression, terminated by a newline:
[5[3(3)4{()]}]
The expression is not balanced.