Q4 smallestValue()
Write Write a function smallestValue() that returns the smallest value
stored in a given non-empty tree. The function accepts a pointer to the root of
the given tree. You should determine the correct function prototype.
int smallestValue(BTNode *node);
Sample cases are given below:
Enter a list of numbers for a Binary Tree, terminated
by any non-digit character:
100001 100095 100088 a
The smallest number in the tree is 100001.
Enter a list of numbers for a Binary Tree, terminated
by any non-digit character:
70 93 1 5 -5 17 8 23 a
The smallest number in the tree is -5.