Q2 removeBSTNode()
Write a function removeBSTNode() that removes a given item from a
Binary Search Tree. The function should return 1 if the item was found and
successfully removed and 0 otherwise. The prototype of the function is given
below.
int removeBSTNode(BTNode **nodePtr, int item);
A sample input and its output is given below:
Enter a list of numbers for a Binary Tree, terminated by any
non-digit character:
31 93 14 70 94 99 23 7 67 25 71 66 -20 0 88 a
The Binary Search Tree:
31
|---14
| |---7
| | |----20
| | | |___0
| |___23
| | |___25
|___93
| |---70
| | |---67
| | | |---66
| | |___71
| | | |___88
| |___94
| | |___99
Enter an integer to be removed from the tree:
93
93 was removed
The Binary Search Tree:
31
|---14
| |---7
| | |----20
| | | |___0
| |___23
| | |___25
|___88
| |---70
| | |---67
| | | |---66
| | |___71
| |___94
| | |___99