Q4 rotateRNode()
Write a function rotateRNode() that rotates right about a given node. The
binary tree will be updated after the function is called. The prototype of the
function is given below.
void rotateRNode(BTNode **node)
A sample input and its output is given below
Enter a list of numbers for a Binary Tree, terminated by any
non-digit character:
70 14 93 7 31 73 99 10 23 45 97 95 a
70
|---14
| |---7
| | |___10
| |___31
| | |---23
| | |___45
|___93
| |---73
| |___99
| | |---97
| | | |---95
Enter an integer to be searched in the tree:
93
Found
about its left child (0) or right child (1): 1
70
|---14
| |---7
| | |___10
| |___31
| | |---23
| | |___45
|___93
| |---73
| |___97
| | |---95
| | |___99
You may also implement the left rotation about the given node.